ColorScale: fix for clicking and dragging outside scale area
The ColorScale widget has padding around the actual scale area that should not
be responsive when clicked or dragged.  When there was no label next to the
scale it wasn't a big deal, because the border was only a couple pixels wide and
if you clicked in the border area and the value was updated, it didn't matter
too much.  However, if there's a label displayed next to the scale, you don't
want the color to be updated (to the minimum value) when the user clicks the
label.  This fix only updates the value if you click inside the scale rectangle.
In addition, it keeps track of whether you started a drag within the scale
rectangle or not.  If you click inside the rectangle and then drag, it will
update the value as long as you have the mouse button held down.  However, if
the first click was in the padding area or label, it won't update the value,
even if you drag it onto the scale rectangle area.