ColorRelation: initial implementation
This is a simple class that models a relationship between colors.  It is
designed such that the 'source' color is monitored for changes, and then the
'destination' color gets updated automatically based on a supplied generator
function.
However, if the destination is modified directly, the relationship class notices
this and updates a local 'offset' parameter that should be added to the
generated color.  This allows the colors to maintain a defined relationship with
eachother while also allowing the parameters of the relationship to be modified
slightly.