Color: Fix operator+ and operator-
Previously, additional and subtraction was done with rgb, but this doesn't
really make much sense for my application.  HSV makes much more sense, so I
re-implemented it to add based on HSV.  I also made is so that the HUE wraps
around if you add something or subtract and get a value beyond the limits.