ColorModel: don't perpetually reconnect to the changed signal
previously we were connecting to the color changed signal in set_color(), but it
doesn't seem like this would be necessary, and might even be resulting in a lot
of wasted resources if we reconnect the exact same handler over and over and
over.  Only connecting to the signal once (in the constructor) seems to work
just fine.