Add ColorEditBox as a new aggregate widget
Previously, I was using a 'ColorEditBox' class in test-ui.cc to make things
easier to display and edit a single color model.  I've ripped this out of the
test and made it an official widgets in libagavewidgets.  The interface isn't
necessary well thought-out yet, but it should suffice for now.