ColorEditBox: Don't set a border_width by default
Previously it was defaulted to 6, but this makes it so that if you want o pack
it right next to another widget with no border, you have to set the border width
explicitly to 0.  It seems better to leave it like other containers and default
to 0 and let the user specify any borders and padding as necessary.