New Approach to Schemes
Previously, the idea of a scheme was that it was a base color and a set of other
colors that satisfied certain relationships.  However that led to problems of
who owned the data when I wanted to display the schemes.  I would have to get
the data (colors) out of the Scheme object and push it into the swatch color
model, etc.

The new concept of a Scheme is a set of relationship between colors.  The scheme
only provides a name and a set of functions that can be used to generate one
color from another one.  Hopefully this makes the problem of data ownership a
little bit easier to deal with.

Disabled the old Scheme tests for now since they no longer compile.  Will
re-enable when I fix the test.