ColorRelation: store local offset as hsv_t so offsets can be negative
previously we stored the local offset as a Color type, but the Color type clamps
the individual components from 0.0 to 1.0, so there can never be a negative
offset (it would get clamped to 0.0).  So instead, we store the offset as an
hsv_t type and do the addition and subtraction that way.
In the meantime I made a hsv_t operator+ and operator- and marked some Color
constructors explicit so that we weren't doing the wrong kind of addition and
subtraction due to implicit conversions.