Color: don't normalize in operator- and operator+
Previously we were normalizing things (e.g. wrapping around for hue) in
operator- and operator+, but this isn't appropriate since the results of a
subtraction may be a negative offset that needs to be retained for the operation
to be reversible (e.g. so that we can later add it and get back to the original
value).

Therefore, the wraparound functionality has been moved to the clamp() function
so that it only gets normalized when we create a Color type from the hsv_t