Swatch: Add ability to set the border width of the swatch
Previously you just just set the border to 'on' or 'off', but now you can set
the width of the border.  This is useful if you want to highlight one particular
swatch with a thicker border or something similar.  If you want to turn the
border 'off', you can just set it to 0.0