Fix up Analogous scheme
Updated the Analogous scheme to generate better schemes.  The implementation is
pretty ugly and there's a lot of duplicate /copy-paste, so I might need to think
about a way to re-factor the scheme functions, but the end result is quite nice.
Also updated the 'test-relations' test to show the analogous scheme