ColorScale: draw RGB sliders differently
Previously the RED, GREEN, and BLUE scales were drawn with a gradient from black
to full red (or green or blue), but this made it difficult to tell what the
resulting color would be when you moved the slider.  This change makes it so
that the scale displays what the resulting color would be if you moved the
slider to a specific point on the scale.

Making this change violated some assumptions that the RED, GREEN, and BLUE
scales don't need to be redrawn if their adjustment is the same as the last
update (since the gradient depends on the values of other channels), so I
changed to cache the last color value (instead of just the last adjustment
value) and update if the color had changed since last update.