ColorSchemeBox: added new convenience class
This class is a convenience wrapper around a set of ColorEditBox widgets.  It
creates 5 color widgets and offers a simplified interface to setting up scheme
relations between them.
Ported the test-relations test to use this new widget and added it to the actual
app (although without setting a particular scheme, so the colors are all
separate at the moment)