ColorSchemeBox: store base color as a variable
Previously, I was referring to the explicit base array element (e.g.
m_edit_boxes[2]) when I wanted to get edit box of the color that serves as the
base of the scheme.  Now I save a pointer to that box in a variable and use that
variable to refer to it.