ColorSchemeBox: add get_colors() to access the underlying color models
Previously, ColorSchemeBox was quite opaque.  You couldn't access the underlying
color models to (for example) use the same model in a different widget.  In
order to show the models in the scheme box and the new color wheel, we needed to
provide access to them so that we can set them in the Color Wheel