ColorWheel: restrict marker size to an integer value
If the marker size is not an integer value, then our calculated diameter for the
widget becomes a non-integer value (since we need to reserve space outside the
color wheel for the marker to 'hang' outside the wheel), which means that our
image surface should be a non-integer value.  But image surfaces may only be
created with integer widths and heights, so when this happened, we occassionally
created slightly-too-small image surfaces, which resulted in some blank space
between the color wheel and the border. Restricting the marker size to an
integer should fix this issue.
Also added a couple logging statements to help figure out what's going on