Create new RgbScaleSet and HsvScaleSet widgets
These widgets encapsulate a set of scales that can be used to modify a color in
a particular colorspace (e.g. rgb or hsv).  You'll usually want to have all
three of these together instead of just being able to modify the red channel of
something, and these classes simplify that.  Also added a base interface
IColorView, which means that you can set a ColorModel that the class 'displays'.
Both of these ScaleSet classes implement this interface.