Swatch: implement IColorView interface
Also changed Swatch to not inherit from DrawingArea, but rather owns a
DrawingArea, to allow more of the implementation to be hidden from the header.
Also ensured that we can only get and set the Color Model to encourage only
modifying the value of the color through the ColorModel class.  This caused some
significant API change, so other files were updated as well, and tests were
updated so they compile again.