IMultiColorView: add new interface for displaying multiple colors
ColorSchemeBox and ColorWheel both implement IMultiColorView since they display
multiple colors at a time.  The ColorSchemeBox no longer has any scheme / color
relation logic -- it's just a dump multi-color swatch/editor display widget.
All of the logic is moved up a level.  The main owner of the ColorModels has
been moved up to ApplicationWindow as well.  This means that we don't have to
get the models from the SchemeBox to give to the ColorWheel anymore -- the
Application owns them and gives them to both the SchemeBox and the ColorWheel.

Overall, it feels like a much better assignment of ownership and responsibilities