Allow viewing and editing of current color with MarkerItem widgets
Now the MarkerItem widgets can be dragged to change the color dynamically.  All
of the colors and marker positions update as expected when the underlying model
is changed and it actually works pretty well already.

In the process of adding this functionality, I also refactored a couple things
to make things a little bit cleaner.

There's a bit of a glitch in the ColorWheel fill pattern -- it doesn't fill
completely to the edges of the item, there are a couple of gaps.  This needs to
be investigated a bit more