Fix issue with incorrect conversion from position to color
This bug was resulting in slightly odd marker dragging behavior, because the
position->color conversion wasn't perfect, but then we set the underlying model,
which then set the position of the marker and it didn't quite match where we had
just dragged the marker to.  So the dragging didn't follow the cursor perfectly.

In the process of fixing this, we no longer do the iterative update position
which updates model which updates position thing.  A side-effect of this is that
we could now drag the marker outside of the valid color wheel area, so we had to
add a drag/drop validation function.