Center the ColorWheel widget in the main application window
The ColorWheel widget should now be centered above the swatches.  It's still a
fixed size, so it needs some work to make it behave nicely when resizing the
window, so for now I've just made it not expand at all when the window resizes.

The ColorWheel markers were also being slightly clipped at the very edge of the
wheel, so I made the wheel radius account for the marker's line width.

Also deleted a debugging print statement