Add ability to parse and save colorsets across sessions with a GMarkup file
The writer is fairly complete and it will save a file (using giomm) when the
ColorSetManager::save() function is called, writing over any existing file with
the same name (perhaps this needs to be refined slightly).

ColorSet required a couple changes to make the API a bit more natural (e.g.
exposing an iterator interface instead of getting a copy of the list of colors).
Also, the ID generation was changed slightly to be based on the color values and
not based on the user and time created, but it may still be incorrect.

The parser is not currently finished, but the infrastructure is there.