Implement a good portion of the saved set parsing
The parser will now load the specified file and parse the name, description, and
ID of the set from the GMarkup file.  I haven't decided on what format to store
the color in (currently I'm saving it as a hex string, but that's not the
long-term solution), so that is not imlplemented in the parser yet.

Improved the error-handling in the file saving code as well.

To parse the saved sets from file, I had to add a set_id() interface to ColorSet